refactor cart and checkout so they share components
