+++ hamza-client/src/app/[countryCode]/(checkout)/checkout/utils/fetch-cart-for-checkout.ts
@@ -1,6 +1,6 @@
 import { enrichLineItems, retrieveCart } from '@modules/cart/actions';
 import { CartWithCheckoutStep } from '@/types/global';
-import { LineItem } from '@medusajs/medusa';
+import { LineItem, Store } from '@medusajs/medusa';
 import { getCheckoutStep } from '@lib/util/get-checkout-step';
 
 /**
@@ -11,7 +11,9 @@ import { getCheckoutStep } from '@lib/util/get-checkout-step';
  * In cart flow we will acquire the cart_id => pass it to the checkout flow... we should pass this as prop since
  * it's a ssc cookie...
  */
-export const fetchCartForCheckout = async (cartId: string): Promise<CartWithCheckoutStep | null> => {
+export const fetchCartForCheckout = async (
+    cartId: string
+): Promise<CartWithCheckoutStep | null> => {
     if (!cartId) return null;
     const cart = await retrieveCart(cartId);
